requirements.txt
PyPDF2==3.0.1
groq==0.36.0
python-dotenv==1.0.1
